Better error message for toContain matcher.
Currently, calling
  expect(null).toContain(42);
will throw an error

minijasminenode/lib/jasmine-1.3.1.js:1142
  return haystack.indexOf(needle) >= 0;
                  ^
TypeError: Cannot call method 'indexOf' of null

Since often the actual is the return value of some function, it can be non-obvious to find the defect. This change will show a more clear error message, pointing the user to the line in the test causing the failure instead of the line in the jasmine code.